+++ packages/graphql-typescript-definitions/src/cli.ts
@@ -1,9 +1,10 @@
 /* eslint no-console: off */
 
-import * as yargs from 'yargs';
 import chalk from 'chalk';
+import yargs from 'yargs';
 
-import {EnumFormat} from './types';
+import {EnumFormat, ExportFormat} from './types';
+
 import {Builder, SchemaBuild, DocumentBuild} from '.';
 
 const argv = yargs
@@ -33,9 +34,13 @@ const argv = yargs
     type: 'boolean',
     describe: 'Add a __typename field to every object type',
   })
+  .option('export-format', {
+    required: false,
+    describe: 'The format to use for values exported from GraphQL documents',
+    choices: [ExportFormat.Document, ExportFormat.Simple],
+  })
   .option('enum-format', {
     required: false,
-    type: 'string',
     describe:
       'The format in which to generate case names for enum types in the schema',
     choices: [
@@ -56,10 +61,10 @@ const argv = yargs
 
 const builder = new Builder({
   cwd: argv.cwd,
-  schemaTypesPath: argv.schemaTypesPath,
-  addTypename: argv.addTypename,
-  enumFormat: argv.enumFormat,
-  customScalars: normalizeCustomScalars(argv.customScalars),
+  schemaTypesPath: argv['schema-types-path'],
+  addTypename: argv['add-typename'],
+  enumFormat: argv['enum-format'],
+  customScalars: normalizeCustomScalars(argv['custom-scalars']),
 });
 
 function normalizeCustomScalars(customScalarOption: string) {