+++ packages/graphql-typescript-definitions/test/schema.test.ts
@@ -217,6 +217,12 @@ describe('printSchema()', () => {
         SIX
       }
 
+      enum Season {
+        FIRST
+        SECOND
+        THIRD
+      }
+
       scalar Date
 
       input InputOne {
@@ -228,12 +234,15 @@ describe('printSchema()', () => {
       input InputTwo {
         name: String
         episode: Episode!
+        season: Season!
         date: Date
       }
     `);
 
     expect(generateSchemaTypes(schema).get('index.ts')).toContain(stripIndent`
+      import { Season } from "./Season";
       import { Episode } from "./Episode";
+      export { Season };
       export { Episode };
       export type Date = string;
       export interface InputOne {
@@ -244,6 +253,7 @@ describe('printSchema()', () => {
       export interface InputTwo {
         name?: string | null;
         episode: Episode;
+        season: Season;
         date?: Date | null;
       }
     `);