+++ packages/graphql-validate-fixtures/src/cli.ts
@@ -1,7 +1,8 @@
 /* eslint no-console: off */
 
 import {relative, basename} from 'path';
-import * as yargs from 'yargs';
+
+import yargs from 'yargs';
 import * as glob from 'glob';
 import chalk from 'chalk';